Revert "Test fixed LV-N plume"
The order of the engines isn't the cause of the plume problem, and this change
causes an additional problem: it assumes that the original engine module uses
LF, but it actually uses LH2 because the LH2NTRsDynamic patch runs beforehand.
The result is that *both* engine modes end up using LH2, though only the first
one (called "LF") actualy works.

Since this change seems to have been a test and it's not necessary to fix the
plume problem, I'm reverting it to fix the propellant problem described above.

This reverts commit e2fc9d150ddae7de501bcca4368b156f350327bd.