Add LH2 conversion patch for Taurus RS-2's radial chemical nozzles
The Taurus RS-2 engine consists of a big nuclear engine in the middle,
surrounded by four small nozzles that may (depending on other MM
patches) be either additional nozzles for the nuke, or separate chemical
engines similar to the stock 24-77 "Twitch" engine.  The LH2NTRsDynamic
patch does a decent job of converting nuclear engines to use LH2 instead
of LF, but if the radial nozzles are configured as chemical LFO engines,
they're left untouched, resulting in a part that uses both LF and LH2
fuel.  Since that's awkward and inconvenient, this patch changes them
from LFO to LH2+O, so that the part needs only LH2 fuel for both its
nuclear and chemical engine components.