De-duplicate generic NTR patches for ModuleEngines and ModuleEnginesFX
The patch is the same regardless of which engine module a part uses, and
it's better not to duplicate the whole patch just to explicitly name the
two different modules when a wildcard can match both.

This raises the possibility that there could be other module types (e.g.
from mods) whose names match the "ModuleEngines*" pattern, but if so,
it's probably appropriate for this patch to apply to them anyway.