Only patch actual nuke engine modules in generic NTR patch
The generic NTR patch excludes multi-mode engines, but a part can also
have multiple engine modules that operate independently (not as
multi-mode alternatives).  Though unusual, it's possible to have both a
nuclear and non-nuclear engine on the same part, or even two
(independent) nuclear engines.

Previously, this patch would modify the part's first engine module, on
the assumption that it's the only one (on a non-multi-mode part).  On a
part with multiple independent engine modules, that could lead to
accidentally patching the non-nuclear engine instead of the nuclear one,
or patching just one of several nuclear engines.  By adding the nuclear
:HAS criteria to the @MODULE block (not just the whole @PART), it'll
correctly patch only the nuclear engine modules even if a non-nuclear
one comes first, and *all* the nuclear engine modules if there's more
than one.