Add nicknames to part tags (and fix a couple tag IDs) (#50)
* Dev (#47)

* fixing but in dynamic NTR patch that caused ISPs to not be ajusted.

* Prepare update to 1.3.1

* Dependency list

* Test fixed LV-N plume

* Revert "Test fixed LV-N plume"

The order of the engines isn't the cause of the plume problem, and this change
causes an additional problem: it assumes that the original engine module uses
LF, but it actually uses LH2 because the LH2NTRsDynamic patch runs beforehand.
The result is that *both* engine modes end up using LH2, though only the first
one (called "LF") actualy works.

Since this change seems to have been a test and it's not necessary to fix the
plume problem, I'm reverting it to fix the propellant problem described above.

This reverts commit e2fc9d150ddae7de501bcca4368b156f350327bd.

* Define separate FX for LV-N's LF and LH2 modes

KSP apparently doesn't like sharing the same effect definitions between several
engine modules; only one of them actually plays the effect.  To get the effects
on both engine modes, they have to be defined twice, with different names.

To avoid duplicating all the effect properties, I've used ModuleManager's
whole-node copy-paste feature to copy the AUDIO and MODULE_MULTI_PARTICLE nodes
from the LH2 effects to the LF ones.

* Fix LV-N engine fxOffset patches

It looks like the LV-N's change from ModuleEngines to ModuleEnginesFX is
supposed to also change the fxOffset attribute, but the patch for the LH2
mode was adding a second fxOffset instead of modifying the first one, and the
LF mode was using the old value instead of the new one.

* Simplify LV-N's LF-mode patch

Instead of duplicating a bunch of properties for the engine's LF engine
module, copy the LH2 module and edit just the properties that need to change.

* Patch LV-N ModuleAlternator to work in both LF and LH2 mode

Without the preferMultiMode flag, the alternator works only when the part's
first engine module (the one for LH2 mode) is running.

* Add comments

The interaction between the LH2NTRsDynamic and hydrogenNTRsSQUAD patches is
non-obvious, and it's the reason why the "first" engine module needs to be the
LH2 one.  Some comments should hopefully make things more clear.

* Versioning bump

* Exclude multi-mode engines from generic LF->LH2 NTR patch

The LH2NTRsDynamic patch looks for engines that seem like
LiquidFuel-burning NTRs and changes them to burn LqdHydrogen instead,
but it assumes that the LiquidFuel-burning engine module is the part's
*only* engine module.  When applied to a multi-mode engine, it can end
up patching the "wrong" engine module (not the LF one), or it might
change an LF/LH2 multi-mode engine to burn LH2 in both modes.

To avoid those problems, it's best to avoid patching multi-mode
engines.  Such parts are specialized enough that they ought to have
part-specific patches (if needed) rather than a one-size-fits-all.

* Prepare 0.4.10

* Added Russian localization (#42)

* Added russian localization

* Fixed minor mistake in ru.cfg

Fixed mistake in Scylla's description

* Prepare 0.4.11

* Add nicknames, fix a #LOC ID

* Fixing a tag ID in a part