Spacetux engine patches (#48)
* Add LH2 conversion patch for LV-2N and LV-4N engine clusters

The LV-2N and LV-4N engine clusters, originally made by Kommitz and
currently distributed as part of the "SpaceTux Industries Recycled
Parts" collection, are basically bigger versions of the stock LV-N
engine, with more thrust but the same Isp.

With Kerbal Atomics installed, all three engines (LV-N, LV-2N, LV-4N)
get the "generic" LH2 conversion patch first, but the LV-N has an
additional patch (in hydrogenNTRsSQUAD.cfg) that changes its Isp and
plume, and adds LF/LH2 multi-mode support.  This new patch makes the
same changes to the LV-2N and LV-4N, to keep the three parts consistent
with each other.

* Add LH2 conversion patch for Taurus RS-2's radial chemical nozzles

The Taurus RS-2 engine consists of a big nuclear engine in the middle,
surrounded by four small nozzles that may (depending on other MM
patches) be either additional nozzles for the nuke, or separate chemical
engines similar to the stock 24-77 "Twitch" engine.  The LH2NTRsDynamic
patch does a decent job of converting nuclear engines to use LH2 instead
of LF, but if the radial nozzles are configured as chemical LFO engines,
they're left untouched, resulting in a part that uses both LF and LH2
fuel.  Since that's awkward and inconvenient, this patch changes them
from LFO to LH2+O, so that the part needs only LH2 fuel for both its
nuclear and chemical engine components.

* Fix Taurus nuke patch vacuum key index

A recent update to the Taurus engine config reversed the order of the
keys in its atmosphereCurve, so that the key for vacuum efficiency is
first instead of last.